+++ app/src/main/java/com/crevado/fr/mywallet/login/data/LoginRequestModel.kt
@@ -1,0 +1,6 @@
+package com.crevado.fr.mywallet.login.data
+
+data class LoginRequestModel(
+    val user: String,
+    val pin: String
+)