+++ app/src/main/java/com/crevado/fr/mywallet/login/presentation/LoginViewState.kt
@@ -1,0 +1,7 @@
+package com.crevado.fr.mywallet.login.presentation
+
+data class LoginViewState(
+    val isLoading: Boolean = false,
+    val data: Any? = null,
+    val error: String = ""
+)