+++ app/src/main/java/com/crevado/fr/mywallet/shared/extensions/KeyboardExtension.kt
@@ -1,0 +1,11 @@
+package com.crevado.fr.mywallet.shared.extensions
+
+import android.app.Activity
+import android.view.inputmethod.InputMethodManager
+
+fun Activity.hideKeyboard() {
+    this.apply {
+        val imm: InputMethodManager = this.getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
+        imm.hideSoftInputFromWindow(this.currentFocus?.windowToken, 0)
+    }
+}