+++ app/src/main/java/com/crevado/fr/mywallet/shared/presentation/MainActivity.kt
@@ -1,8 +1,11 @@
-package com.crevado.fr.mywallet
+package com.crevado.fr.mywallet.shared.presentation
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.crevado.fr.mywallet.R
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)