+++ app/src/main/java/com/crevado/fr/mywallet/user_login/data/UserLoginRequestModel.kt
@@ -1,6 +1,0 @@
-package com.crevado.fr.mywallet.user_login.data
-
-data class UserLoginRequestModel(
-    val user: String,
-    val pin: String
-)