+++ app/src/main/java/com/crevado/fr/mywallet/utils/BundleKey.kt
@@ -1,0 +1,11 @@
+package com.crevado.fr.mywallet.utils
+
+object BundleKey {
+    const val KEY_USER_ID = "KEY_USER_ID"
+    const val KEY_USER_NAME = "KEY_USER_NAME"
+    const val KEY_USER_WALLET = "KEY_USER_WALLET"
+    const val KEY_USER_EMAIL = "KEY_USER_EMAIL"
+    const val KEY_USER_IMAGE_URL = "KEY_USER_IMAGE_URL"
+    const val KEY_BALANCE = "KEY_BALANCE"
+    const val KEY_CURRENCY = "KEY_CURRENCY"
+}