+++ app/src/main/java/com/crevado/fr/mywallet/utils/Constant.kt
@@ -1,0 +1,5 @@
+package com.crevado.fr.mywallet.utils
+
+object Constant {
+    const val BASE_URL = "https://wind-assessment-api.vercel.app/api/"
+}